[FIX] html_editor*: move wrapInlinesInBlocks to dom plugin *: mail, website
Before this commit: the wrapInlinesInBlocks is in utils.

After this commit: now wrapInlinesInBlocks is part of dom plugin.

There are two cases wrapInlinesInBlocks is used out of editor or during
instantiating the editor:
1. used in formatDefaultBodyForFullComposer in mail, which should be
removed because we will call wrapInlinesInBlocks when initialize the
editor, calling it beforehand is potentially wrong when
allowInlineAtRoot is true.
2. used in initElementForEdition, since we call this.preparePlugins()
before initElementForEdition, we have access to wrapInlinesInBlocks in
the dom plugin.

task-5109662
