+++ addons/html_editor/static/src/utils/dom.js
@@ -8,13 +8,9 @@ import {
     previousLeaf,
 } from "./dom_info";
 import { callbacksForCursorUpdate } from "./selection";
-import { isEmptyBlock, isPhrasingContent } from "../utils/dom_info";
+import { isEmptyBlock } from "../utils/dom_info";
 import { childNodes, descendants } from "./dom_traversal";
 import { childNodeIndex, DIRECTIONS, nodeSize } from "./position";
-import {
-    baseContainerGlobalSelector,
-    createBaseContainer,
-} from "@html_editor/utils/base_container";
 
 /** @typedef {import("@html_editor/core/selection_plugin").Cursors} Cursors */
 
@@ -36,87 +32,6 @@ export function makeContentsInline(node) {
             currentNode = unwrapContents(currentNode)[0];
         } else {
             currentNode = currentNode.nextSibling;
-        }
-    }
-}
-
-/**
- * Wrap inline children nodes in Blocks, optionally updating cursors for
- * later selection restore. A paragraph is used for phrasing node, and a div
- * is used otherwise.
- *
- * @param {HTMLElement} element - block element
- * @param {Cursors} [cursors]
- */
-export function wrapInlinesInBlocks(
-    element,
-    { baseContainerNodeName = "P", cursors = { update: () => {} } } = {}
-) {
-    // Helpers to manipulate preserving selection.
-    const wrapInBlock = (node, cursors) => {
-        const block = isPhrasingContent(node)
-            ? createBaseContainer(baseContainerNodeName, node.ownerDocument)
-            : node.ownerDocument.createElement("DIV");
-        cursors.update(callbacksForCursorUpdate.append(block, node));
-        cursors.update(callbacksForCursorUpdate.before(node, block));
-        if (node.nextSibling) {
-            const sibling = node.nextSibling;
-            node.remove();
-            sibling.before(block);
-        } else {
-            const parent = node.parentElement;
-            node.remove();
-            parent.append(block);
-        }
-        block.append(node);
-        return block;
-    };
-    const appendToCurrentBlock = (currentBlock, node, cursors) => {
-        if (currentBlock.matches(baseContainerGlobalSelector) && !isPhrasingContent(node)) {
-            const block = currentBlock.ownerDocument.createElement("DIV");
-            cursors.update(callbacksForCursorUpdate.before(currentBlock, block));
-            currentBlock.before(block);
-            for (const child of childNodes(currentBlock)) {
-                cursors.update(callbacksForCursorUpdate.append(block, child));
-                block.append(child);
-            }
-            cursors.update(callbacksForCursorUpdate.remove(currentBlock));
-            currentBlock.remove();
-            currentBlock = block;
-        }
-        cursors.update(callbacksForCursorUpdate.append(currentBlock, node));
-        currentBlock.append(node);
-        return currentBlock;
-    };
-    const removeNode = (node, cursors) => {
-        cursors.update(callbacksForCursorUpdate.remove(node));
-        node.remove();
-    };
-
-    const children = childNodes(element);
-    const visibleNodes = new Set(children.filter(isVisible));
-
-    let currentBlock;
-    let shouldBreakLine = true;
-    for (const node of children) {
-        if (isBlock(node)) {
-            shouldBreakLine = true;
-        } else if (!visibleNodes.has(node)) {
-            removeNode(node, cursors);
-        } else if (node.nodeName === "BR") {
-            if (shouldBreakLine) {
-                wrapInBlock(node, cursors);
-            } else {
-                // BR preceded by inline content: discard it and make sure
-                // next inline goes in a new Block
-                removeNode(node, cursors);
-                shouldBreakLine = true;
-            }
-        } else if (shouldBreakLine) {
-            currentBlock = wrapInBlock(node, cursors);
-            shouldBreakLine = false;
-        } else {
-            currentBlock = appendToCurrentBlock(currentBlock, node, cursors);
         }
     }
 }
